add tenant limiter, move event subscription to webhook

+++ src/Migrations/Version20230922194158.php
@@ -27,12 +27,14 @@ final class Version20230922194158 extends AbstractMigration
             $schema->dropTable('fusio_event_trigger');
         }
 
-        $eventResponseTable = $schema->getTable('fusio_event_response');
-        if (!$eventResponseTable->hasColumn('body')) {
-            $eventResponseTable->removeForeignKey('event_response_trigger_id');
-            $eventResponseTable->dropColumn('trigger_id');
-            $eventResponseTable->dropColumn('error');
-            $eventResponseTable->addColumn('body', 'text', ['notnull' => false]);
+        if ($schema->hasTable('fusio_event_response')) {
+            $eventResponseTable = $schema->getTable('fusio_event_response');
+            if (!$eventResponseTable->hasColumn('body')) {
+                $eventResponseTable->removeForeignKey('event_response_trigger_id');
+                $eventResponseTable->dropColumn('trigger_id');
+                $eventResponseTable->dropColumn('error');
+                $eventResponseTable->addColumn('body', 'text', ['notnull' => false]);
+            }
         }
     }
